+++ src/main/java/com/servlet/Exceptions/CreateObjectException.java
@@ -1,0 +1,7 @@
+package com.servlet.Exceptions;
+
+public class CreateObjectException extends Exception{
+        public CreateObjectException(String errorMessage) {
+            super(errorMessage);
+        }
+}
