+++ src/main/java/com/servlet/Exceptions/DoesNotExist.java
@@ -1,0 +1,7 @@
+package com.servlet.Exceptions;
+
+public class DoesNotExist extends Exception {
+    public DoesNotExist(String errorMessage) {
+        super(errorMessage);
+    }
+}
