+++ src/main/java/com/servlet/Exceptions/MultipleObjectsReturned.java
@@ -1,0 +1,7 @@
+package com.servlet.Exceptions;
+
+public class MultipleObjectsReturned extends Exception{
+    public MultipleObjectsReturned(String errorMessage) {
+        super(errorMessage);
+    }
+}
