+++ src/main/java/com/servlet/Exceptions/ParseException.java
@@ -1,0 +1,7 @@
+package com.servlet.Exceptions;
+
+public class ParseException extends Exception{
+    public ParseException(String errorMessage) {
+        super(errorMessage);
+    }
+}
