+++ src/main/java/com/servlet/Models/BaseModel.java
@@ -1,0 +1,10 @@
+package com.servlet.Models;
+
+
+public class BaseModel {
+    public Long id;
+
+    public BaseModel(Long id){
+        this.id = id;
+    }
+}
