+++ src/main/java/com/servlet/ObjectFactory/ObjectFactory.java
@@ -1,0 +1,13 @@
+package com.servlet.ObjectFactory;
+
+import com.servlet.Exceptions.ParseException;
+
+import java.util.Map;
+
+public class ObjectFactory {
+    public Object create(Map<String, Object> rawObj) throws NoSuchFieldException, ClassNotFoundException, ParseException { return null; }
+
+    public Object create(Long id, Map<String, Object> rawObj){ return null; }
+
+    public Object create(Long id) { return null; }
+}
