+++ src/main/java/com/servlet/ObjectFactory/TweetFactory.java
@@ -1,0 +1,21 @@
+package com.servlet.ObjectFactory;
+
+import com.servlet.Exceptions.ParseException;
+import com.servlet.Models.Tweet;
+
+import java.util.Map;
+
+public class TweetFactory extends ObjectFactory{
+    @Override
+    public Object create(Map<String, Object> rawObj) throws NoSuchFieldException, ClassNotFoundException, ParseException {
+        return new Tweet(rawObj);
+    }
+
+    @Override
+    public Object create(Long id, Map<String, Object> rawObj){
+        return new Tweet(id, rawObj);
+    }
+
+    @Override
+    public Object create(Long id){ return new Tweet(id); }
+}
