+++ src/main/java/com/servlet/ObjectFactory/UserFactory.java
@@ -1,0 +1,19 @@
+package com.servlet.ObjectFactory;
+import com.servlet.Models.User;
+
+import java.util.Map;
+
+public class UserFactory extends ObjectFactory{
+    @Override
+    public Object create(Map<String, Object> rawObj){
+        return new User(rawObj);
+    }
+
+    @Override
+    public Object create(Long id, Map<String, Object> rawObj){
+        return new User(id, rawObj);
+    }
+
+    @Override
+    public Object create(Long id) { return new User(id); }
+}
