+++ src/main/java/com/servlet/Serializers/ListSerializer.java
@@ -1,0 +1,9 @@
+package com.servlet.Serializers;
+
+import org.json.JSONObject;
+
+import java.util.List;
+
+public abstract class ListSerializer {
+    abstract public List<JSONObject> data();
+}
