+++ src/main/java/com/servlet/Settings/Settings.java
@@ -1,0 +1,5 @@
+package com.servlet.Settings;
+
+public class Settings {
+    public static boolean DEBUG = true;
+}
