+++ src/main/resources/img/log-out.png
