+++ src/main/resources/js/posts.js
