--- src/main/webapp/WEB-INF/classes/Exceptions/CreateObjectException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class CreateObjectException extends Exception{
-        public CreateObjectException(String errorMessage) {
-            super(errorMessage);
-        }
-}
