--- src/main/webapp/WEB-INF/classes/Exceptions/DoesNotExist.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class DoesNotExist extends Exception {
-    public DoesNotExist(String errorMessage) {
-        super(errorMessage);
-    }
-}
