--- src/main/webapp/WEB-INF/classes/Exceptions/MultipleObjectsReturned.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class MultipleObjectsReturned extends Exception{
-    public MultipleObjectsReturned(String errorMessage) {
-        super(errorMessage);
-    }
-}
