--- src/main/webapp/WEB-INF/classes/Exceptions/ParseException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class ParseException extends Exception{
-    public ParseException(String errorMessage) {
-        super(errorMessage);
-    }
-}
