--- src/main/webapp/WEB-INF/classes/Models/BaseModel.java
@@ -1,10 +1,0 @@
-package Models;
-
-
-public class BaseModel {
-    public Long id;
-
-    public BaseModel(Long id){
-        this.id = id;
-    }
-}
