--- src/main/webapp/WEB-INF/classes/ObjectFactory/ObjectFactory.java
@@ -1,13 +1,0 @@
-package ObjectFactory;
-
-import Exceptions.ParseException;
-
-import java.util.Map;
-
-public class ObjectFactory {
-    public Object create(Map<String, Object> rawObj) throws NoSuchFieldException, ClassNotFoundException, ParseException { return null; }
-
-    public Object create(Long id, Map<String, Object> rawObj){ return null; }
-
-    public Object create(Long id) { return null; }
-}
