--- src/main/webapp/WEB-INF/classes/Serializers/ListSerializer.java
@@ -1,9 +1,0 @@
-package Serializers;
-
-import org.json.JSONObject;
-
-import java.util.List;
-
-public abstract class ListSerializer {
-    abstract public List<JSONObject> data();
-}
