--- src/main/webapp/WEB-INF/classes/Servlets/FileUploadServlet.java
@@ -1,54 +1,0 @@
-package Servlets;
-
-import java.io.File;
-import java.io.IOException;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.MultipartConfig;
-import javax.servlet.http.Part;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@MultipartConfig
-public class FileUploadServlet extends HttpServlet {
-    private static final String UPLOAD_DIR = "uploads";
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) {
-        String applicationPath = request.getServletContext().getRealPath("");
-        String uploadPath = applicationPath + File.separator + UPLOAD_DIR;
-
-        File uploadDir = new File(uploadPath);
-        if (!uploadDir.exists()) {
-            uploadDir.mkdirs();
-        }
-
-        System.out.println("Fileb being uploaded to: " + uploadPath);
-        try {
-            String filePath = null;
-            for (Part part : request.getParts()) {
-                filePath = getFilePath(part);
-                part.write(uploadPath + File.separator + filePath);
-            }
-
-            System.out.println("File uploaded to: " + uploadPath);
-            request.getRequestDispatcher("./").forward(request, response);
-        } catch (IOException e) {
-            response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-        } catch (ServletException e) {
-            response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    private String getFilePath(Part part) {
-        String contentDisp = part.getHeader("content-disposition");
-        String[] tokens = contentDisp.split(";");
-        for (String token : tokens) {
-            if (token.trim().startsWith("filename")) {
-                return token.substring(token.indexOf("=") + 2, token.length() - 1);
-            }
-        }
-        return "";
-    }
-}
