--- src/main/webapp/WEB-INF/classes/Settings/Settings.java
@@ -1,5 +1,0 @@
-package Settings;
-
-public class Settings {
-    public static boolean DEBUG = true;
-}
