+++ src/main/webapp/index.html
@@ -4,9 +4,9 @@
 <head>
     <meta charset="UTF-8">
     <title>Avoci</title>
-    <link rel="shortcut icon" type="image" href="resources/img/logo-ico.png">
+    <link rel="shortcut icon" type="image" href="../resources/img/logo-ico.png">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="resources/css/styles.css">
+    <link rel="stylesheet" href="../resources/css/styles.css">
 </head>
 
 <body>
@@ -15,7 +15,7 @@
             <div class="left-block">
                 <div class="header-logo">
                     <a href="#" title="Homepage">
-                        <img id="logo" src="resources/img/logo.png" title="Homepage" alt="Logo">
+                        <img id="logo" src="../resources/img/logo.png" title="Homepage" alt="Logo">
                     </a>
                 </div>
                 <div>
@@ -28,12 +28,12 @@
                 </div>
                 <div class="new-post">
                     <a href="#" class="new-post" title="New post">
-                        <img class="header-icons" src="resources/img/add-post.png" title="Profile" alt="Add new post">
+                        <img class="header-icons" src="../resources/img/add-post.png" title="Profile" alt="Add new post">
                     </a>
                 </div>
                 <div class="log-out">
                     <button class="log-out-btn">
-                        <img class="header-icons" src="resources/img/log-out.png" title="Log out" alt="Log out">
+                        <img class="header-icons" src="../resources/img/log-out.png" title="Log out" alt="Log out">
                     </button>
                 </div>
             </div>
@@ -43,7 +43,7 @@
         <div class="make-new-post-area">
             <form>
                 <div class="user-photo-column">
-                    <img class="user-photo" src="resources/img/user-photo.jpg" alt="User photo">
+                    <img class="user-photo" src="../resources/img/user-photo.jpg" alt="User photo">
                 </div>
 
                 <div class="container-column">
@@ -81,7 +81,7 @@
         </div>
         <article class="user-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/user-photo.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/user-photo.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -125,7 +125,7 @@
         </article>
         <article class="other-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/planet.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/planet.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -155,7 +155,7 @@
         </article>
         <article class="other-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/comp.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/comp.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -188,7 +188,7 @@
         </article>
         <article class="user-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/user-photo.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/user-photo.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -233,7 +233,7 @@
         </article>
         <article class="other-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/comp.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/comp.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -264,7 +264,7 @@
         </article>
         <article class="user-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/user-photo.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/user-photo.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -307,7 +307,7 @@
         </article>
         <article class="other-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/planet.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/planet.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -339,7 +339,7 @@
         </article>
         <article class="user-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/user-photo.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/user-photo.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -383,7 +383,7 @@
         </article>
         <article class="other-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/planet.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/planet.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -413,7 +413,7 @@
         </article>
         <article class="other-post">
             <div class="user-photo-column">
-                <img class="user-photo" src="resources/img/comp.jpg" alt="User photo">
+                <img class="user-photo" src="../resources/img/comp.jpg" alt="User photo">
             </div>
             <div class="container-column">
                 <div class="post-info">
@@ -453,7 +453,7 @@
         </div>
     </footer>
 </body>
-<script src="resources/js/posts.js"></script>
-<script src="resources/js/index.js"></script>
+<script src="../resources/js/posts.js"></script>
+<script src="../resources/js/index.js"></script>
 
 </html>
