--- src/main/webapp/resources/css/styles.css
@@ -1,226 +1,0 @@
-body {
-    padding: 0;
-    margin: 0;
-    color: black;
-}
-body, html {
-    height: 100%;
-}
-header{
-    width: 100%;
-    position: fixed;
-    background-color: white;
-    left: 0;
-    top: 0;
-    box-shadow: gray 0 2px 10px;
-}
-.header-content {
-    width: 92%;
-    margin: auto;
-    padding: 22px 0;
-    height: 30px;
-    display: flex;
-}
-.left-block, .right-block {
-    display: flex;
-}
-.right-block{
-    margin-left: auto;
-}
-.header-username{
-   color: black;
-    font-family: "Times New Roman", serif;
-    font-style: italic;
-    margin-top: 8px;
-}
-.header-icons{
-    width: 30px;
-    padding-left: 5px;
-    vertical-align: middle;
-}
-.search{
-    border-radius: 50px;
-    border: 1px gray solid;
-    height: 27px;
-    margin-top: 5px;
-    outline: none;
-}
-#logo{
-    width: 180px;
-    padding-right: 20px;
-    margin-top: -17px;
-}
-.posts-container {
-    max-width: 1300px;
-    margin: 60px auto 60px;
-    padding-top: 50px;
-    width: 90%;
-    overflow: hidden;
-
-}
-.new-post{
-    margin-top: auto;
-    margin-left: 3%;
-}
-.make-new-post-area{
-    background-color: white;
-    border-radius: 30px;
-    padding: 0.5rem;
-    width: 97%;
-    border: 1px solid #92d36e;
-}
-.user-photo-column, .container-column{
-    display: inline-block;
-}
-.user-photo-column{
-    width: 9%;
-    vertical-align: top;
-}
-.container-column{
-    width: 90%;
-    margin-top: auto;
-}
-.user-photo{
-    border-radius: 50%;
-    width: 80px;
-}
-.post-text{
-    margin-top: 1%;
-}
-.new-post-textarea{
-    resize: none;
-    overflow: hidden;
-    background-color: white;
-    outline: none;
-    width: 100%;
-    border: 1px solid white;
-    color: gray;
-    font-family: "Arial", serif;
-    font-style: italic;
-}
-.add-new-post{
-    width: 8%;
-    margin-left: auto;
-}
-.action-area{
-    width: 100%;
-    display: flex;
-}
-.action-container{
-    margin-left: auto;
-    display: flex;
-}
-.posts-text{
-    width: 85%;
-    margin-top: -3%;
-}
-.add-post-btn, .load-more-btn{
-    border-radius: 50px;
-    width: 100%;
-    border: none;
-    background-color: #72bb53;
-    color: white;
-    text-align: center;
-}
-.load-more-btn{
-    width: 20%;
-    margin-top: 20px;
-    height: 2em;
-    margin-left: 39%;
-}
-.filter-area{
-    background-color: white;
-    border-radius: 30px;
-    padding: 0.5rem;
-    width: 97%;
-    border: 1px solid #92d36e;
-    margin-top: 20px;
-}
-fieldset{
-    background-color: white;
-    outline: none;
-    width: 97%;
-    border: 1px solid white;
-    color: gray;
-    font-family: "Arial", serif;
-    font-style: italic;
-}
-.filter-content{
-    width: 96%;
-}
-.filter-container{
-    padding-bottom: -10px;
-    display: none;
-}
-.filter-textarea{
-    resize: none;
-    overflow: hidden;
-    background-color: white;
-    outline: none;
-    width: 30%;
-    border: 1px solid lightgray;
-    color: gray;
-    font-family: "Arial", serif;
-    font-style: italic;
-    border-radius: 5px;
-    height: 15px;
-    margin-right: 15px;
-}
-.user-post, .other-post{
-    margin-top: 20px;
-    background-color: white;
-    border-radius: 30px;
-    padding: 0.5rem;
-    width: 97%;
-    border: 1px solid #92d36e;
-}
-.post-info{
-    display: flex;
-    padding-bottom: -5%;
-}
-.date-and-time{
-    margin-left: auto;
-    font-size: medium;
-}
-.add-photo{
-    width: 25px;
-}
-.username{
-    color: black;
-    font-style: italic;
-}
-.like{
-    margin-left: auto;
-}
-.username-in-post{
-    margin-top: -1%;
-}
-.like-btn, .change-post-btn, .delete-post-btn, .log-out-btn, .btn-apply{
-    background: white;
-    border: none;
-    outline: none;
-}
-footer .footer-content {
-    width: 92%;
-    max-width: 1300px;
-    margin: auto;
-    height: 36px;
-}
-footer{
-    left: 0;
-    bottom: 0;
-    width: 100%;
-    flex: 0 0 auto;
-    background-color: white;
-    text-align: center;
-    color: gray;
-    border-top: 1px darkgray solid;
-}
-@media only screen and (max-width: 1000px) {
-    .column-1, .column-2{
-        display: grid;
-    }
-    .user-post, .other-post, .make-new-post-area, .filter-area{
-        width: 95%;
-    }
-}
